# packages/frontend/src/redux/schema/api-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/transactions/{id}": {
      "get": {
        "operationId": "Transactions_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Return transaction by id",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TransactionDto" }
              }
            }
          }
        },
        "tags": ["Transactions"]
      }
    },
    "/api/transaction-categories": {
      "post": {
        "operationId": "TransactionCategories_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransactionCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "properties": { "payload": { "type": "string" } } }
              }
            }
          }
        },
        "tags": ["Transaction categories"]
      },
      "get": {
        "operationId": "TransactionCategories_findAllByUser",
        "parameters": [
          {
            "name": "visibilityType",
            "required": false,
            "in": "query",
            "schema": { "$ref": "#/components/schemas/visibilityType" }
          }
        ],
        "responses": {
          "200": {
            "description": "Return transaction category by id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransactionCategoryDto"
                  }
                }
              }
            }
          }
        },
        "tags": ["Transaction categories"]
      }
    },
    "/api/transaction-categories/{id}": {
      "get": {
        "operationId": "TransactionCategories_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionCategoryDto"
                }
              }
            }
          }
        },
        "tags": ["Transaction categories"]
      },
      "patch": {
        "operationId": "TransactionCategories_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTransactionCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionCategoryDto"
                }
              }
            }
          }
        },
        "tags": ["Transaction categories"]
      },
      "delete": {
        "operationId": "TransactionCategories_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["Transaction categories"]
      }
    },
    "/api/transaction-categories/{id}/summary": {
      "get": {
        "operationId": "TransactionCategories_getCategorySummary",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryMonthlySummaryDto"
                  }
                }
              }
            }
          }
        },
        "tags": ["Transaction categories"]
      }
    },
...
    "/api/incomes": {
      "get": {
        "operationId": "Incomes_findAllByUser",
        "parameters": [
          {
            "name": "month",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "year",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "accountTypes",
            "required": false,
            "in": "query",
            "schema": { "type": "array", "items": { "type": "string" } }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/PaginationDto" },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/IncomeDto" }
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": ["Incomes"]
      },
      "post": {
        "operationId": "Incomes_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateIncomeDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/IncomeDto" }
              }
            }
          }
        },
        "tags": ["Incomes"]
      }
    },
    "/api/incomes/monthly-summaries": {
      "get": {
        "operationId": "Incomes_findMonthlySummariesByuser",
        "parameters": [
          {
            "name": "month",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "year",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "accountTypes",
            "required": false,
            "in": "query",
            "schema": { "type": "array", "items": { "type": "string" } }
          },
          {
            "name": "transactionCategories",
            "required": false,
            "in": "query",
            "schema": { "type": "array", "items": { "type": "string" } }
          },
          {
            "name": "parentTransactionCategory",
            "required": false,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransactionMonthSummaryDto"
                  }
                }
              }
            }
          }
        },
        "tags": ["Incomes"]
      }
    },
    "/api/incomes/{id}": {
      "get": {
        "operationId": "Incomes_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Return transaction by id",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/IncomeDto" }
              }
            }
          }
        },
        "tags": ["Incomes"]
      },
      "patch": {
        "operationId": "Incomes_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateIncomeDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/IncomeDto" }
              }
            }
          }
        },
        "tags": ["Incomes"]
      },
      "delete": {
        "operationId": "Incomes_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["Incomes"]
      }
    },
...
  },
  "info": {
    "title": "Financer",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateTransactionCategoryMappingWithoutTransactionDto": {
        "type": "object",
        "properties": {
          "description": { "type": "string" },
          "categoryId": { "type": "string" },
          "amount": { "type": "number" }
        },
        "required": ["categoryId", "amount"]
      },
      "TransactionDto": {
        "type": "object",
        "properties": {
          "createdAt": { "format": "date-time", "type": "string" },
          "updatedAt": { "format": "date-time", "type": "string" },
          "id": { "type": "string" },
          "amount": { "type": "number" },
          "description": { "type": "string" },
          "date": { "format": "date-time", "type": "string" },
          "userId": { "type": "string" },
          "fromAccount": { "type": "string" },
          "toAccount": { "type": "string" },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateTransactionCategoryMappingWithoutTransactionDto"
            }
          }
        },
        "required": [
          "createdAt",
          "updatedAt",
          "id",
          "amount",
          "description",
          "date",
          "userId",
          "fromAccount",
          "toAccount",
          "categories"
        ]
      },
...
      "TransactionType": {
        "type": "string",
        "enum": ["INCOME", "EXPENSE", "TRANSFER"]
      },
      "TransactionCategoryDto": {
        "type": "object",
        "properties": {
          "createdAt": { "format": "date-time", "type": "string" },
          "updatedAt": { "format": "date-time", "type": "string" },
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "name": { "type": "string" },
          "visibility": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/TransactionType" }
          },
          "deleted": { "type": "boolean" },
          "parentCategoryId": { "type": "string", "nullable": true }
        },
        "required": [
          "createdAt",
          "updatedAt",
          "id",
          "userId",
          "name",
          "visibility",
          "deleted",
          "parentCategoryId"
        ]
      },
...
      "PaginationDto": {
        "type": "object",
        "properties": {
          "currentPage": { "type": "number" },
          "totalPageCount": { "type": "number" },
          "totalRowCount": { "type": "number" },
          "limit": { "type": "number" },
          "hasPreviousPage": { "type": "boolean" },
          "hasNextPage": { "type": "boolean" }
        },
        "required": [
          "currentPage",
          "totalPageCount",
          "totalRowCount",
          "limit",
          "hasPreviousPage",
          "hasNextPage"
        ]
      },
...
      "TransactionMonthSummaryIdDto": {
        "type": "object",
        "properties": {
          "year": { "type": "number" },
          "month": { "type": "number" }
        },
        "required": ["year", "month"]
      },
      "TransactionMonthSummaryDto": {
        "type": "object",
        "properties": {
          "id": { "$ref": "#/components/schemas/TransactionMonthSummaryIdDto" },
          "count": { "type": "number" },
          "totalCount": { "type": "number" },
          "incomesCount": { "type": "number" },
          "expensesCount": { "type": "number" },
          "transferCount": { "type": "number" },
          "totalAmount": { "type": "number" },
          "incomeAmount": { "type": "number" },
          "expenseAmount": { "type": "number" },
          "transferAmount": { "type": "number" }
        },
        "required": [
          "id",
          "count",
          "totalCount",
          "incomesCount",
          "expensesCount",
          "transferCount",
          "totalAmount",
          "incomeAmount",
          "expenseAmount",
          "transferAmount"
        ]
      },
...
      "CategoryMonthlySummaryDto": {
        "type": "object",
        "properties": {
          "_id": { "type": "object" },
          "count": { "type": "number" },
          "totalAmount": { "type": "number" }
        },
        "required": ["_id", "count", "totalAmount"]
      },
      "CreateTransactionCategoryDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "visibility": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/TransactionType" }
          },
          "parentCategoryId": { "type": "string", "nullable": true }
        },
        "required": ["name", "visibility", "parentCategoryId"]
      },
      "visibilityType": {
        "type": "string",
        "enum": ["INCOME", "EXPENSE", "TRANSFER"]
      },
      "UpdateTransactionCategoryDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "visibility": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/TransactionType" }
          },
          "parentCategoryId": { "type": "string", "nullable": true }
        }
      },
...
      "IncomeDto": {
        "type": "object",
        "properties": {
          "createdAt": { "format": "date-time", "type": "string" },
          "updatedAt": { "format": "date-time", "type": "string" },
          "id": { "type": "string" },
          "amount": { "type": "number" },
          "description": { "type": "string" },
          "date": { "format": "date-time", "type": "string" },
          "userId": { "type": "string" },
          "toAccount": { "type": "string" },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateTransactionCategoryMappingWithoutTransactionDto"
            }
          }
        },
        "required": [
          "createdAt",
          "updatedAt",
          "id",
          "amount",
          "description",
          "date",
          "userId",
          "toAccount",
          "categories"
        ]
      },
      "CreateIncomeDto": {
        "type": "object",
        "properties": {
          "amount": { "type": "number" },
          "description": { "type": "string" },
          "date": { "format": "date-time", "type": "string" },
          "toAccount": { "type": "string" },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateTransactionCategoryMappingWithoutTransactionDto"
            }
          }
        },
        "required": ["amount", "description", "date", "toAccount", "categories"]
      },
      "UpdateIncomeDto": {
        "type": "object",
        "properties": {
          "amount": { "type": "number" },
          "description": { "type": "string" },
          "date": { "format": "date-time", "type": "string" },
          "toAccount": { "type": "string" },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateTransactionCategoryMappingWithoutTransactionDto"
            }
          }
        }
      },
...
    }
  }
}